xsktable initial altering logic
